fix duplicate constants in e2e tests (#174)
While working with the codebase, I noticed some duplicated constants in
the e2e tests. This PR cleans up the test code by using the existing
TEST_PRIVATE_KEY and DEFAULT_DELEGATION_ADDRESS constants instead of
duplicating values.
This makes the code more maintainable - if we need to change these
values in the future, we'll only need to update them in one place. Also
fixed the error handling to be consistent with the rest of the codebase.
Small changes, but should make the tests cleaner and easier to maintain.